move and cleanup document applicables
