+++ document_applicables/rules/util.py
